try moving gorilla out of utils
