+++ lib/web/ws_io.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package utils
+package web
 
 import (
 	"github.com/gorilla/websocket"
